+++ 09/src/academy/pocu/comp2500samples/w09/magician/Water.java
@@ -1,36 +1,36 @@
-package academy.pocu.comp2500samples.w09.elementalist;
+package academy.pocu.comp2500samples.w09.magician;
 
 import java.time.Duration;
 import java.time.OffsetDateTime;
 import java.time.ZoneOffset;
 
-public class Earth extends Attunement {
-    private final int ELITE_SKILL_COOLDOWN_TIME_SEC = 100;
+public class Water extends Attunement {
+    private final int ELITE_SKILL_COOLDOWN_TIME_SEC = 75;
 
-    public Earth(Elementalist elementalist) {
-        super(elementalist);
+    public Water(Magician magician) {
+        super(magician);
     }
 
     public void onEntry() {
-        System.out.println("As heavy as earth!");
+        System.out.println("Water! The giver of life!");
     }
 
     public void attack() {
-        System.out.println("Stone edge!");
+        System.out.println("Water gun!");
     }
 
     public void useEliteSkill(OffsetDateTime dateTime) {
         OffsetDateTime now = OffsetDateTime.now(ZoneOffset.UTC);
 
         if (now.compareTo(dateTime.plus(Duration.ofSeconds(ELITE_SKILL_COOLDOWN_TIME_SEC))) > 0) {
-            System.out.println("Feel the wrath of Mother Earth!");
-            System.out.println("Earthquake!");
+            System.out.println("Alright, let's surf!");
+            System.out.println("Poseidon's fury! Tsunami~~~!!");
         } else {
             System.out.println("I can't use this yet!");
         }
     }
 
     public void onDeath() {
-        System.out.println("Now I rest in Mother Earth...");
+        System.out.println("Help! I can't swim!");
     }
 }
