+++ 09/src/academy/pocu/comp2500samples/w09/objectequality/Program.java
@@ -1,4 +1,4 @@
-package academy.pocu.comp2500samples.w09.objectequals;
+package academy.pocu.comp2500samples.w09.objectequality;
 
 import java.util.HashSet;
